Keep going back and forth on this, but I think I want the value/input modes in own files.
