+++ src/controller/requests/InitialVisualizerRequest.js
@@ -1,6 +1,8 @@
 import VisualizerRequest from "./VisualizerRequest";
 import { AvailableGames } from "../../model/enums/AvailableGames";
-import { FilterRequest, DropdownItem, InputModes, ValueModes } from "./FilterRequest";
+import InputModes from "../../model/enums/InputModes";
+import ValueModes from "../../model/enums/ValueModes";
+import { FilterRequest, DropdownItem } from "./FilterRequest";
 import { InitialVisualizerModel } from "../../model/visualizations/InitialVisualizerModel";
 
 /**
